added new ui

+++ Assets/Scripts/AttackSystem/AttackProvider.cs
@@ -3,13 +3,13 @@ using UnityEngine.Events;
 
 public class AttackProvider : MonoBehaviour, IAttackProvider
 {
-    [SerializeField] private float m_Damage = 0;
+    [SerializeField] private int m_Damage = 0;
     [Space]
     public UnityEvent<IAttackProvider, IAttackReciever> OnSuccessAttack = new();
 
     public bool Active { get; set; } = true;
 
-    public float Damage { get => m_Damage; set => m_Damage = value; }
+    public int Damage { get => m_Damage; set => m_Damage = value; }
 
     public void OnSuccessHit(IAttackReciever reciever)
     {
