+++ Assets/Scripts/AttackSystem/Base/IAttackProvider.cs
@@ -2,7 +2,7 @@ public interface IAttackProvider
 {
     public bool Active { get; }
 
-    public float Damage { get; }
+    public int Damage { get; }
 
     void OnSuccessHit(IAttackReciever reciever);
 }
