+++ Assets/Scripts/Stats/IStats.cs
@@ -4,5 +4,5 @@
     public float InvincibleTime { get; }
     public float MoveSpeed { get; }
     public float AttackSpeed { get; }
-    public float AttackDamage { get; }
+    public int AttackDamage { get; }
 }
