+++ Assets/Scripts/Stats/ModificatedStats.cs
@@ -3,12 +3,12 @@
     private IStats m_SourceStats;
     private float m_MoveSpeedModificator;
     private float m_AttackSpeedModificator;
-    private float m_AttackDamageModificator;
+    private int m_AttackDamageModificator;
 
     public ModificatedStats(IStats sourceStats,
         float moveSpeedModificator = 0,
         float attackSpeedModificator = 0,
-        float attackDamageModificator = 0
+        int attackDamageModificator = 0
         )
     {
         m_SourceStats = sourceStats;
@@ -21,5 +21,5 @@
     public float InvincibleTime => m_SourceStats.InvincibleTime;
     public float MoveSpeed => m_SourceStats.MoveSpeed + m_MoveSpeedModificator;
     public float AttackSpeed => m_SourceStats.AttackSpeed + m_AttackSpeedModificator;
-    public float AttackDamage => m_SourceStats.AttackDamage + m_AttackDamageModificator;
+    public int AttackDamage => m_SourceStats.AttackDamage + m_AttackDamageModificator;
 }
